pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8.0,<3.9"
numpy = "^1.22.2"
replit = "^3.2.4"
"discord.py" = "^1.7.3"
Command = "^0.1.0"
Flask = "^2.2.2"
# ChatterBot = "^1.1.0" for l8r. kills the repl.
pylint = "^2.15.4"
black = "^22.10.0"
flake8 = "^5.0.4"
disnake = "^2.6.0"
toolz = "^0.12.0"
ChatterBot = "1.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"